+++ hapro/CHANGELOG.md
@@ -1,5 +1,13 @@
 <!-- https://developers.home-assistant.io/docs/add-ons/presentation#keeping-a-changelog -->
 
+## v1.1 21/01/2026
+### Ownership Transfer
+- The ownership of the HaPro addon has been transferred to Ludero
+
+### Fixes
+- Fixed an issue where homeassistant would get corrupted on enabling systemmonitor statistics
+
+
 ## v1.0 17/01/2025
 # File Distribution
 - Added the system for file distribution